friend declaration fix to overcome required encapsulation along with moving the ara::core detail into utility header
